Files

Description

???

Documentation

type Content = String

???

onlyFiles :: String -> [String] -> [Name]

Remove . and .. from dir contents and extend path of files

shortName :: Name -> Name

Shorten file name, dropping dir names

initObjects :: [Name] -> [Content] -> IO [Object]

Get objects from DB

catchErrors :: [(Name, Maybe Location)] -> IO [Object]

Find configs with errors, remove them from list and write msgs about it

getCoords :: Content -> Maybe Location

Read file content

makeTuples :: [String] -> [(Maybe Int, Maybe Int)]

Make tuples of ints

readInts :: [(Maybe Int, Maybe Int)] -> Maybe Location

Try to read ints

checkInts :: (Maybe Int, Maybe Int) -> Maybe Location -> Maybe Location

Check if all coords have proper format

saveWorld :: World -> IO World

Save current uni as config

saveUni :: Universe -> IO ()

Create new file with unique name and save config

makeConfig :: Universe -> String

Make config file out of matrix

getAlive :: [(Int, Int, Cell)] -> String

Convert coords of alive cells to string