Game

Description

???

Documentation

data Cell

???

Constructors

Dead
???

Alive
???

Half
???

data State

Coords - last mouse position

Constructors

Generator
???

Iterator
???

CfgMenu Int
???

ObjMenu Int (Maybe Coords)
???

type Universe = Matrix Cell

???

type Coords = (Int, Int)

???

type Location = [Coords]

???

type Name = String

???

type MenuItem = Int

???

data World

???

Constructors

World

Fields

universe :: Universe

???

state :: State

???

obj :: Objects

???

cfg :: Configs

???

selected :: MenuItem

???

pic :: [Picture]

???

age :: Integer

???

data Object

???

Constructors

Object

Fields

name :: Name

???

coords :: Location

???

data Objects

???

Constructors

Objects

Fields

oList :: [Object]

???

oNum :: Int

???

data Configs

???

Constructors

Configs

Fields

cList :: [Object]

???

cNum :: Int

???

size :: Int

Constant field size

defState :: Universe

Default state

Functions to convert object and config coords to universe

loadObject :: Universe -> Object -> Maybe Coords -> Universe

???

loadObjectAux :: Coords -> Coords -> Coords -> Universe -> Universe

???

loadConfig :: Object -> Universe

???

Functions to convert cell states

halfToAlive :: Universe -> Universe

???

convert :: Cell -> Cell

???

inverseCell :: Cell -> Cell

???

isAlive :: Cell -> Bool

Check if the cell is alive

getNeighbours :: Universe -> Coords -> [Cell]

Get Moore neighbourhood of a cell as a list of neighbouring cells

stepCellAux :: Cell -> [Cell] -> Cell

Updates the cell according to the number of its neighbours

stepCell :: Universe -> Coords -> Cell

Update function for cell

stepUniverseAux :: Universe -> Universe -> Coords -> Universe

Auxillary function to update the universe cell by cell

stepUniverse :: Universe -> Universe

Universe updater